rocksock_operation(): do not use select() for SSL connections
when we use SSL, we can't rely on select() not timing out because
calls to other SSL IO functions (such as SSL_peek,read,write) may
already have cleared the socket and buffered whatever data was
available. in that case it's sufficient to just call the read/write
funcs.
whether the timeout has any effect tho needs to be investigated.