fix rocksock_peek for both ssl backends
the semantics are now roughly equivalent to rocksock_peek on a
non-ssl socket. when there's data to read however, the internal
calls to SSL_peek will cause blocking, this shouldnt hurt much however
as you would do a blocking call to read the available data anyway
immediately after getting the peek result.
in case there's no data on the socket, it will not block and return
immediately.