treat orderly shutdown as error
this makes it easier to deal with disconnects (for example when
automatic reconnect is desired), as you can treat them as every
connection error without having a separate codepath for the
"orderly shutdown" special condition.

a library using rocksock that does peek and read in the same
function would have to return both values if the error handling
code was in the caller, as the value 0 in the recv_bytes argument
could either be interpreted as a disconnect, or a peek when no
data was waiting. now it can reserve the meaning of 0 for its own
purposes.