rocksock_readline: change semantics: trailing \n is removed
this has the advantage that we dont need one additional byte
to store the zero terminator. so if the protocol limits line
lengths for example to 512, you can pass a 512 byte buffer
and any message will fit into it and is properly terminated.
otherwise you'd have to pass an odd-sized 513 byte buffer,
and in most cases would strip off the trailing \n anyway.

the size returned in the size_t pointer now is effectively
strlen(line) or bytes received (including line delimiter) -1.