API change: put the onus to allocate proxy array on the user
this has one big advantage:
we don't have to hardcode the max amount of proxies into the
library itself, so it's possible to ship the precompiled library
while still allowing the user to freely choose the max amount
of proxies he wants to use.

another advantage is that the user can allocate the exact amount
of memory he needs for the number of proxies he's going to use,
without wasting any ram for unneeded slots.

the last advantage is that we can refactor the rs_proxy array
to contain the string buffers for all the fields, so we can also
get rid of the NO_STRDUP ifdef and dependent code.