merge all callsites for unexpected error
reduces codesize since we can replace a function call and a jump
with just a jump.
disadvantage is that the line number information for the specific
error site gets lost. however that info is only useful for
debugging, and when doing so one can reason from where the jump
happened.