add support for RcB2 build system
as the libulz include files, when installed systemwide, live in an "ulz" prefix,
e.g. /include/ulz/foo.h, but not when they're in their native source tree, an
additional USE_LIBULZ_SYS ifdef was added for the former case. when compiling
with RcB2 and the wish to use USE_LIBULZ, you need to specify the path to its
include dir with a -I switch added to C/CPPFLAGS.

micclient/server have not been adapted so far because they have pending changes.